Fix libp2p helper failing due to logging
Problem: when log message from libp2p helper contains a
"$" character, Daemon crashes.

Solution: make the Daemon not crash, but merely log an error
of logging.

Commentary: long-term we probably do not want to process logs
from Helper via Daemon, but for now it's enough to just stop
the node from crashing. Logs must not ever be a reason for crashing.